+++ tests/ResponseListTest.php
@@ -18,6 +18,7 @@
 namespace Phpoaipmh;
 
 use Phpoaipmh\Fixture\ClientStub;
+use Phpoaipmh\HttpAdapter\HttpAdapterInterface;
 use PHPUnit_Framework_TestCase;
 
 /**
@@ -144,7 +145,7 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
     public function testGetRecordsInCollectionReturnsValueIfSuppliedByEndpoint()
     {
         $obj = $this->getSampleMultiPageResponseList();
-        $this->assertEquals(733, $obj->getTotalRecordsInCollection());
+        $this->assertEquals(733, $obj->getTotalRecordCount());
     }
 
     // ---------------------------------------------------------------
@@ -156,7 +157,19 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
         $obj = new RecordIterator($this->getMockClient($output), 'ListRecords');
 
         // The single page records do not supply a number
-        $this->assertNull($obj->getTotalRecordsInCollection());
+        $this->assertNull($obj->getTotalRecordCount());
+    }
+
+    // ---------------------------------------------------------------
+
+    public function testGetRecordsWithNamespaces()
+    {
+        //Single page sample file contains 162 results in a valid ListRecords response
+        $output = $this->generateSampleXML(array('NamespacedSample.xml'));
+        $obj = new RecordIterator($this->getMockClient($output), 'ListRecords');
+
+        // The single page records do not supply a number
+        $this->assertEquals(2402263, $obj->getTotalRecordCount());
     }
 
     // ----------------------------------------------------------------
@@ -205,8 +218,9 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
     /**
      * Get a mock client
      *
-     * @param array $toReturn  Array of values to return for consecutive calls (send one for same every time)
-     * @return \Phpoaipmh\HttpAdapter\HttpAdapterInterface
+     * @param array $toReturn Array of values to return for consecutive calls (send one for same every time)
+     * @return HttpAdapterInterface|ClientStub
+     * @throws \Exception
      */
     protected function getMockClient($toReturn = array())
     {